optimized_cloning/requirements.txt
bs4==0.0.1
GitPython==3.1.37